+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.6.5
 
+* JsonSchema: Update Hydra `@context` property possible types
 * Filter validation: Fix issue in Required filter validator with dot notation (#4221)
 * OpenAPI: Fix notice/warning for `response` without `content` in the `openapi_context` (#4210)
 * Serializer: Convert internal error to HTTP 400 in Ramsey uuid denormalization from invalid body string (#4200)